+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Bug Fixes
 
+- Fixed a bug when transitioning Jira issues from 'Start work', which doesn't refresh the issues panels
 - Added sorting for status transitions based on their workflow order
 
 ## What's new in 3.8.6